Move Prepare and Finalize into CliqueBlockProducer
